t/xml/charref-1.dat

#data
&#x20;<a/>
#errors
1;6;text outside of root element;;;m
#document
| " "
| <{} a>

#data
&#32;<a/>
#errors
1;5;text outside of root element;;;m
#document
| " "
| <{} a>

#data
 &#x20;<a/>
#errors
1;7;text outside of root element;;;m
#document
| " "
| <{} a>

#data
&#x20; <a/>
#errors
1;6;text outside of root element;;;m
#document
| "  "
| <{} a>

#data
<a/>
&#x20;
#errors
2;5;text outside of root element;;;m
#document
| <{} a>
| " "

#data
<a/>&#32;
#errors
1;8;text outside of root element;;;m
#document
| <{} a>
| " "

#data
<!DOCTYPE a>&#x20;<a/>
#errors
1;18;text outside of root element;;;m
#document
| <!DOCTYPE a>
| " "
| <{} a>

#data
<!DOCTYPE a> &#x20;<a/>
#errors
1;19;text outside of root element;;;m
#document
| <!DOCTYPE a>
| " "
| <{} a>

#data
<!DOCTYPE a>&#x20; <a/>
#errors
1;18;text outside of root element;;;m
#document
| <!DOCTYPE a>
| "  "
| <{} a>

#data
<a>&#x20;</a>
#errors
#document
| <{} a>
|   " "

#data
<a>&#X20;</a>
#errors
1;6;uppercase hcro;;;m
#document
| <{} a>
|   " "

#data
<a>&abcdefghijklmnopqrstuvwxyzabcdefghijklmnopqrstuvwxyz;</a>
#errors
1;4;entity not declared;;abcdefghijklmnopqrstuvwxyzabcdefghijklmnopqrstuvwxyz;;m
#document
| <{} a>
|   "&abcdefghijklmnopqrstuvwxyzabcdefghijklmnopqrstuvwxyz;"

#data
<a>&#x0;</a>
#errors
1;8;invalid character reference;U+0000;;m
#document escaped
| <{} a>
|   "\uFFFD"

#data
<a>&#0;</a>
#errors
1;7;invalid character reference;U+0000;;m
#document escaped
| <{} a>
|   "\uFFFD"

#data
<a>&#x1;</a>
#errors
error
#document escaped
| <{} a>
|   "\u0001"

#data
<a>&#01;</a>
#errors
error
#document escaped
| <{} a>
|   "\u0001"

#data
<a>&#xd800;</a>
#errors
1;11;invalid character reference;U+D800;;m
#document escaped
| <{} a>
|   "\uFFFD"

#data
<a>&#55296;</a>
#errors
1;11;invalid character reference;U+D800;;m
#document escaped
| <{} a>
|   "\uFFFD"

#data
<a>&#xdfff;</a>
#errors
1;11;invalid character reference;U+DFFF;;m
#document escaped
| <{} a>
|   "\uFFFD"

#data
<a>&#57343;</a>
#errors
1;11;invalid character reference;U+DFFF;;m
#document escaped
| <{} a>
|   "\uFFFD"

#data
<a>&#x0D;</a>
#errors
#document escaped
| <{} a>
|   "\u000D"

#data
<a>&#13;</a>
#errors
#document escaped
| <{} a>
|   "\u000D"
